slay of the day

+++ final-project/insta-baddie/lib/actions/likes.js
@@ -4,30 +4,22 @@ import { useRouter } from "next/navigation";
 
 export default function Likes({ post, addOptimisticPost}) {
     const router = useRouter();
-    const handleLikes = async () => {
+    
+  const handleLikes = async () => {
     const supabase = createClientComponentClient();
-    const {
-      data: { user },
-    } = await supabase.auth.getUser();
+    const {data: { user },} = await supabase.auth.getUser();
+
     if (user) {
       if (post.user_has_liked_post) {
-        addOptimisticPost({
-            ...post,
-            likes: post.likes-1,
-            user_has_liked_post: !post.user_has_liked_post,
-        })
-        await supabase
-          .from("likes")
-          .delete()
-          .match({ user_id: user.id, post_id: post.id });
+        addOptimisticPost({post,likes: post.likes-1,user_has_liked_post: !post.user_has_liked_post,})
+        await supabase.from("likes").delete().match({ user_id: user.id, post_id: post.id });
       } else {
-        await supabase
-          .from("likes")
-          .insert({ user_id: user.id, post_id: post.id });
+        await supabase.from("likes").insert({ user_id: user.id, post_id: post.id });
       }
       router.refresh();
     }
   };
+
   return (
     <button onClick={handleLikes} className="group flex items-center">
       <svg
